requirements.txt
torch>=2.0
transformers>=4.30
peft>=0.6
opencv-python-headless
scikit-learn
pyyaml
tqdm
huggingface_hub
